requirements.txt
# ── Core API Framework ─────────────────────────────────────────────────────
# fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── HTTP client (async) ────────────────────────────────────────────────────
httpx==0.27.0

# ── ML / AI Models ─────────────────────────────────────────────────────────
transformers==4.41.2
torch==2.3.0                  # CPU-only; auto-selected by pip on non-GPU machines
sentence-transformers==3.0.1

# ── Image Processing ───────────────────────────────────────────────────────
Pillow==10.3.0
imagehash==4.3.1

# ── NLP ────────────────────────────────────────────────────────────────────
spacy==3.7.4

# ── Numerics ───────────────────────────────────────────────────────────────
numpy==1.26.4
scikit-learn==1.5.0

# ── Config ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
